Added support for os.execute and io.popen
New permission handler permexec allows controlled execution of external
programs (as subprocesses of the host program) with sanitized
environment.
External program execution requests are used by both os.execute and
io.popen. These requests work based on a combined hash of the contents
of the executable (NOT the name) and all arguments.